\name{bp}
\alias{bp}
\title{
Add unit "bp" to the number
}
\description{
Add unit "bp" to the number
}
\usage{
bp(x)
}
\arguments{

  \item{x}{a numeric vector. }

}
\details{
It adds a new \code{bp} class to the vector.
}
\value{
A same vector as input
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
bp(10)
bp(10.1)
}
