\name{cr_concatenate}
\alias{cr_concatenate}
\title{
Concatenate two CR objects
}
\description{
Concatenate two CR objects
}
\usage{
cr_concatenate(cr1, cr2)
}
\arguments{

  \item{cr1}{correlated regions}
  \item{cr2}{correlated regions}

}
\value{
A concatenated CR object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
