\name{cr_coverage_on_genes}
\alias{cr_coverage_on_genes}
\title{
CR coverage on genes
}
\description{
CR coverage on genes
}
\usage{
cr_coverage_on_genes(cr, txdb)
}
\arguments{

  \item{cr}{correlated regions}
  \item{txdb}{transcriptome annotation which was used in \code{\link{correlated_regions}}}

}
\examples{
# There is no example
NULL
}
