\name{cr_enriched_heatmap_at_gene}
\alias{cr_enriched_heatmap_at_gene}
\title{
Visualizing enrichment for epigenomic signals at gene body
}
\description{
Visualizing enrichment for epigenomic signals at gene body
}
\usage{
cr_enriched_heatmap_at_gene(cr, txdb, expr, cgi, K = 1, marks = NULL, type = "neg", extend = 5000,
    expr_annotation)
}
\arguments{

  \item{cr}{-cr}
  \item{txdb}{-txdb}
  \item{expr}{-expr}
  \item{cgi}{-cgi}
  \item{K}{-K}
  \item{marks}{-marks}
  \item{type}{-type}
  \item{extend}{-extend}
  \item{expr_annotation}{}

}
\details{

}
\value{
no value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
