\name{cr_enriched_heatmap_at_tss}
\alias{cr_enriched_heatmap_at_tss}
\title{
Visualizing enrichment for epigenomic signals at TSS
}
\description{
Visualizing enrichment for epigenomic signals at TSS
}
\usage{
cr_enriched_heatmap_at_tss(cr, txdb, expr, cgi,
    fdr_cutoff = 0.05, meth_diff_cutoff = 0.1, marks = NULL, type = "neg", extend = c(5000, 10000),
    expr_annotation)
}
\arguments{

  \item{cr}{-cr}
  \item{txdb}{-txdb}
  \item{expr}{-expr}
  \item{cgi}{-cgi}
  \item{fdr_cutoff}{-fdr_cutoff}
  \item{meth_diff_cutoff}{-meth_diff_cutoff}
  \item{marks}{-marks}
  \item{type}{-type}
  \item{extend}{-extend}
  \item{expr_annotation}{}

}
\details{

}
\value{
no value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
