\name{[.chromatin_states_transition_matrix}
\alias{[.chromatin_states_transition_matrix}
\title{
Subset chromatin_states_transition_matrix class object
}
\description{
Subset chromatin_states_transition_matrix class object
}
\usage{
\method{[}{chromatin_states_transition_matrix}(x, i, j, drop = FALSE)
}
\arguments{

  \item{x}{a \code{chromatin_states_transition_matrix} class object}
  \item{i}{index of rows}
  \item{j}{index of columns}
  \item{drop}{whether degenerate the matrix}

}
\value{
a \code{chromatin_states_transition_matrix} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
