\name{[.subgroup_specific_genomic_regions}
\alias{[.subgroup_specific_genomic_regions}
\title{
Subset the subgroup_specific_genomic_regions class object
}
\description{
Subset the subgroup_specific_genomic_regions class object
}
\usage{
\method{[}{subgroup_specific_genomic_regions}(x, i)
}
\arguments{

  \item{x}{a \code{\link{subgroup_specific_genomic_regions}} object}
  \item{i}{index}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
