\name{initialize_project_directory}
\alias{initialize_project_directory}
\title{
Initialize project directories
}
\description{
Initialize project directories
}
\usage{
initialize_project_directory(project_dir)
}
\arguments{

  \item{project_dir}{path of the project directory}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
