\name{plot_subgroup_specificity_heatmap}
\alias{plot_subgroup_specificity_heatmap}
\title{
Heatmap for visualizing subgroup specific genomic regions
}
\description{
Heatmap for visualizing subgroup specific genomic regions
}
\usage{
plot_subgroup_specificity_heatmap(gr_list, genomic_features = NULL)
}
\arguments{

  \item{gr_list}{a list of \code{\link[GenomicRanges]{GRanges}} objects which is generated by \code{\link{subgroup_specific_genomic_regions}}}
  \item{genomic_features}{Genomic features that are used for annotating to regions in \code{gr_list}, it should be a list of \code{\link[GenomicRanges]{GRanges}} objects}

}
\details{
Columns are clustered within each subgroup.
}
\value{
A \code{\link[ComplexHeatmap]{HeatmapList-class}} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
