\name{print.bp}
\alias{print.bp}
\title{
Print bp class objects
}
\description{
Print bp class objects
}
\usage{
\method{print}{bp}(x, ...)
}
\arguments{

  \item{x}{\code{\link{bp}} class object}
  \item{...}{other arguments}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
