\name{print.chromatin_states_transition_matrix}
\alias{print.chromatin_states_transition_matrix}
\title{
Print chromatin_states_transition_matrix class object
}
\description{
Print chromatin_states_transition_matrix class object
}
\usage{
\method{print}{chromatin_states_transition_matrix}(x, ...)
}
\arguments{

  \item{x}{a \code{chromatin_states_transition_matrix} class object}
  \item{...}{additional arguments}

}
\value{
no value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
