\name{print.methylation_hooks}
\alias{print.methylation_hooks}
\title{
Print the methylation_hooks object
}
\description{
Print the methylation_hooks object
}
\usage{
\method{print}{methylation_hooks}(x, ...)
}
\arguments{

  \item{x}{a \code{\link{methylation_hooks}} objects}
  \item{...}{additional arguments}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
