\name{print.subgroup_specific_genomic_regions}
\alias{print.subgroup_specific_genomic_regions}
\title{
Print subgroup_specific_genomic_regions class object
}
\description{
Print subgroup_specific_genomic_regions class object
}
\usage{
\method{print}{subgroup_specific_genomic_regions}(x, ...)
}
\arguments{

  \item{x}{a \code{\link{subgroup_specific_genomic_regions}} class object}
  \item{...}{additional arguments}

}
\value{
no value is returned
}
\author{
ZUguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
