\name{reduce_cr_gap_test}
\alias{reduce_cr_gap_test}
\title{
Test a proper gap value for cr merging
}
\description{
Test a proper gap value for cr merging
}
\usage{
reduce_cr_gap_test(cr)
}
\arguments{

  \item{cr}{correlated regions generated by \code{\link{filter_correlated_regions}}}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
