\name{reduce_david_results}
\alias{reduce_david_results}
\title{
reduce david results
}
\description{
reduce david results
}
\usage{
reduce_david_results(tb, fdr_cutoff = 0.05, hit_cutoff = 5)
}
\arguments{

  \item{tb}{object from \code{\link{submit_to_david}}}
  \item{fdr_cutoff}{cutoff for fdr}
  \item{hit_cutoff}{cutoff for number of genes in a term}

}
\examples{
# There is no example
NULL

}
