\name{set_counter}
\alias{set_counter}
\title{
Set a counter which represents the percentage finished in a loop
}
\description{
Set a counter which represents the percentage finished in a loop
}
\usage{
set_counter(n, fmt = "\%s")
}
\arguments{

  \item{n}{number of iteration in a loop.}
  \item{fmt}{format of the message, pass to \code{\link[base]{sprintf}}. "\%s" will be replaced with the percent string generated by the counter.}

}
\details{
The counter function should be executed in every iteration in the loop. The message only shows in interactive R session.
}
\value{
A counter function.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
counter = set_counter(1000)
for(i in 1:1000) {counter()}
counter = set_counter(1000, fmt = "processing \%s")
for(i in 1:1000) {counter()}
}
