\name{state_names}
\alias{state_names}
\title{
Simply return names of chromatin states
}
\description{
Simply return names of chromatin states
}
\usage{
state_names(x)
}
\arguments{

  \item{x}{a \code{chromatin_states_transition_matrix} object returned from \code{\link{make_transition_matrix_from_chromHMM}}}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
