\name{state_names<-}
\alias{state_names<-}
\title{
Change chromatin state names
}
\description{
Change chromatin state names
}
\usage{
state_names<-(x, value)
}
\arguments{

  \item{x}{a \code{chromatin_states_transition_matrix} object returned from \code{\link{make_transition_matrix_from_chromHMM}}}
  \item{value}{new chromatin state names}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
