\name{t.chromatin_states_transition_matrix}
\alias{t.chromatin_states_transition_matrix}
\title{
Transpose the chromatin transition matrix
}
\description{
Transpose the chromatin transition matrix
}
\usage{
\method{t}{chromatin_states_transition_matrix}(x)
}
\arguments{

  \item{x}{a \code{chromatin_states_transition_matrix} class object}

}
\value{
a \code{chromatin_states_transition_matrix} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
