\name{cr_enriched_heatmap_at_gene}
\alias{cr_enriched_heatmap_at_gene}
\title{
Visualizing enrichment for epigenomic signals at gene body
}
\description{
Visualizing enrichment for epigenomic signals at gene body
}
\usage{
cr_enriched_heatmap_at_gene(cr, txdb, expr, cgi, K = 1, marks = NULL,
    extend = 5000, expr_ha)
}
\arguments{

  \item{cr}{correalted regions}
  \item{txdb}{transcriptome annotation which was used in \code{\link{correlated_regions}}}
  \item{expr}{expression matrix which was used in \code{\link{correlated_regions}}}
  \item{cgi}{CpG island, a \code{\link[GenomicRanges]{GRanges}} object}
  \item{K}{which k-means cluster which is generated by \code{\link{cr_enriched_heatmap}}}
  \item{marks}{names of histone marks, should be supported in \code{\link{chipseq_hooks}}}
  \item{extend}{base pairs extended to upstream and downstream}
  \item{expr_ha}{a \code{\link[ComplexHeatmap]{HeatmapAnnotation}} class object.It is used for the expression heatmap}

}
\details{
There are several heatmaps visualize various signals enriched at gene body

There are following heatmaps:

\itemize{
  \item heatmap for gene expression
  \item a point plot showing gene length
  \item If \code{cr} is returned form \code{\link{cr_enriched_heatmap}}, there is a one column heatmap which shows the k-means cluters genes belong to
  \item heatmap for CGI enrichment at TSS
  \item heatmap for correlation between methylation and gene expression
  \item heatmap for mean methylation
  \item heatmap for metnylation difference
  \item heatmap for correlation, mean signal and signal difference for histone marks
}

If there are more than 12 heatmaps, they will be put into two pages.

Heatmaps are split into three sub-clusters by k-means clustering on the mean methylation matrix.
If there are two subgroups in all samples, each subcluster are split by high expression/low expression
in subgroup 1. In each high expression/low expression, rows are split by the k-means clusters calculated
in \code{\link{cr_enriched_heatmap}}. Finally, rows are clustered by mean methylation matrix.
}
\value{
no value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
