\name{print.bp}
\alias{print.bp}
\title{
Print the bp object
}
\description{
Print the bp object
}
\usage{
\method{print}{bp}(x, ...)
}
\arguments{

  \item{x}{a \code{\link{bp}} class object}
  \item{...}{additional arguments}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
